Made timer an object that stores id and boolean
The reason for this is because if the component updates
multiple times before the timer has called its deexcitation
function (e.g. if the user is spamming buttons), then
the deexcitation function will be called multiple times.
Storing the boolean, started, will enable me to clear the
timer in between componentDidUpdate calls.